Added Data, modifed Controller for better read. modified:   class/ScenarioSAE.class modified:   class/controller/Controller$1.class modified:   class/controller/Controller.class modified:   class/dao/CommuneService.class modified:   class/data/Aeroport.class modified:   class/data/Annee.class modified:   class/data/Commune.class modified:   class/view/MainPage$1$1.class modified:   class/view/MainPage$1.class modified:   class/view/MainPage$2$1.class modified:   class/view/MainPage$2.class modified:   class/view/MainPage.class modified:   src/ScenarioSAE.java modified:   src/controller/Controller.java modified:   src/model/dao/CommuneService.java modified:   src/model/data/Aeroport.java modified:   src/model/data/Annee.java modified:   src/model/data/Commune.java modified:   src/view/MainPage.java

+++ src/model/dao/CommuneService.java
@@ -45,12 +45,12 @@ public class CommuneService {
                 // Récupération des données annuelles
                 int nbMaison = resultSet.getInt("nbMaison");
                 int nbAppart = resultSet.getInt("nbAppart");
-                float prixMoyen = resultSet.getFloat("prixMoyen");
-                float prixM2Moyen = resultSet.getFloat("prixM2Moyen");
-                float surfaceMoyenne = resultSet.getFloat("SurfaceMoy");
-                float depensesCulturellesTotales = resultSet.getFloat("depensesCulturellesTotales");
-                float budgetTotal = resultSet.getFloat("budgetTotal");
-                float population = resultSet.getFloat("population");
+                double prixMoyen = resultSet.getDouble("prixMoyen");
+                double prixM2Moyen = resultSet.getDouble("prixM2Moyen");
+                double surfaceMoyenne = resultSet.getDouble("SurfaceMoy");
+                double depensesCulturellesTotales = resultSet.getDouble("depensesCulturellesTotales");
+                double budgetTotal = resultSet.getDouble("budgetTotal");
+                int population = resultSet.getInt("population");
 
                 try {
                     Commune commune = new Commune(idCommune, nomCommune, nbMaison, nbAppart, prixMoyen,
